requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
Pillow>=9.0.0
concrete-ml>=1.3.0
openfhe>=1.1.1
